Establishing credentials in one file.
There is now only a single file that must be altered in order to change the source of the service account credentials.

There are two cases for the method of retrieving these credentials:
1. in development branches off of stage
2. in the cloudapps deployment

In a development branch, the credentials .json will be in the root directory, and loaded from filepath.

In development, the credentials must be extracted from the environment variable secrets.

To switch between these two methods depending on the branch's needs, the method is commented out manually.